# 6.x Performance Improvements