--- a/src/Elasticsearch.Net/Responses/ElasticsearchResponse.cs
+++ b/src/Elasticsearch.Net/Responses/ElasticsearchResponse.cs
@@ -6,7 +6,7 @@ using System.Text;
 namespace Elasticsearch.Net
 {
 	/// <summary>
-	/// A response from elasticsearch including details about the request/response life cycle
+	/// A response from Elasticsearch including details about the request/response life cycle
 	/// </summary>
 	public abstract class ElasticsearchResponseBase : IApiCallDetails, IElasticsearchResponse
 	{
@@ -47,7 +47,7 @@ namespace Elasticsearch.Net
 	}
 
 	/// <summary>
-	/// A response from elasticsearch including details about the request/response life cycle. Base class for the built in low level response
+	/// A response from Elasticsearch including details about the request/response life cycle. Base class for the built in low level response
 	/// types: <see cref="StringResponse"/> <see cref="BytesResponse"/> and <see cref="DynamicResponse"/>
 	/// </summary>
 	public abstract class ElasticsearchResponse<T> : ElasticsearchResponseBase