… src/Elasticsearch.Net/Providers/RecyclableMemoryStream.cs
… src/Elasticsearch.Net/Transport/Transport.cs
… src/Nest/Aggregations/AggregateJsonConverter.cs
… src/Nest/CommonAbstractions/LazyDocument/LazyDocument.cs
… src/Nest/CommonAbstractions/LazyDocument/LazyDocumentJsonConverter.cs
… src/Nest/CommonAbstractions/SerializationBehavior/ElasticContractResolver.cs
… src/Nest/CommonAbstractions/SerializationBehavior/GenericJsonConverters/ErrorCauseJsonConverter.cs
… src/Nest/CommonAbstractions/SerializationBehavior/GenericJsonConverters/GenericProxyRequestConverterBase.cs
… src/Nest/CommonAbstractions/SerializationBehavior/InternalSerializer.cs
… src/Nest/CommonAbstractions/SerializationBehavior/JTokenExtensions.cs
… src/Nest/CommonAbstractions/SerializationBehavior/JsonReaderExtensions.cs
… src/Nest/CommonAbstractions/SerializationBehavior/SourceConverter.cs
… src/Nest/CommonAbstractions/SerializationBehavior/StatefulDeserialization/JsonConverterPiggyBackState.cs
… src/Nest/CommonAbstractions/Union/UnionJsonConverter.cs
… src/Nest/CommonOptions/Attributes/AlternativeEnumMemberAttribute.cs
… src/Nest/Document/Multiple/Bulk/BulkResponseItem/BulkResponseItemBase.cs
… src/Nest/QueryDsl/Abstractions/Container/QueryContainer-Dsl.cs
… src/Nest/QueryDsl/Abstractions/Query/BoolQueryOrExtensions.cs
… src/Nest/QueryDsl/Abstractions/Query/QueryBase.cs
… src/Nest/QueryDsl/VariableFields.cs
… src/Nest/Search/Search/SearchRequest.cs
… src/Nest/Search/SearchShards/SearchShardsRequest.cs
… src/Nest/Search/Suggesters/SuggestDictionary.cs
… src/Serializers/Nest.JsonNetSerializer/ConnectionSettingsAwareSerializerBase.Customization.cs
… src/Serializers/Nest.JsonNetSerializer/Converters/HandleNestTypesOnSourceJsonConverter.cs
… src/Serializers/Nest.JsonNetSerializer/JTokenExtensions.cs
… src/Serializers/Nest.JsonNetSerializer/JsonNetSerializer.cs
… src/Serializers/Nest.JsonNetSerializer/JsonReaderExtensions.cs
… src/Tests/ClientConcepts/ConnectionPooling/Dispose/ResponseBuilderDisposeTests.cs
… src/Tests/ClientConcepts/HighLevel/Inference/FieldInference.doc.cs