@Mpdreamz 2018-06-11
merged and ported to `master`